feat: upload profile picture

+++ src/accounts/accountsController.js
@@ -5,6 +5,7 @@ const { validateImageInput, deleteFile } = require('./helper');
 
 async function uploadProfilePicture(req, res, next) {
   try {
+    console.log('req.file', req.file);
     const error = validateImageInput(req.body, req.file);
 
     if (error) {
@@ -24,7 +25,9 @@ async function uploadProfilePicture(req, res, next) {
 
     new responseHandler(res, account, 200, RESPONSE_MESSAGE.SUCCESS);
   } catch (error) {
-    deleteFile(req.file.path);
+    if (req.file){
+      deleteFile(req.file.path);
+    }
     return next(error);
   }
 }
